requirements.txt
gunicorn==20.1.0
alembic==1.12.1

Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
greenlet==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
SQLAlchemy==2.0.23
Werkzeug==2.3.6
flask-script==2.0.5
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
pandas==2.0.2
pytest-flask==1.3.0
pytest==7.4.0
factory_boy==2.12.0
faker